+++ index.html
@@ -177,7 +177,8 @@ npm
       </div>
     </section>
   </template>
-  <script src="./js/main.js"></script>
+  <script src="./js/main.js" type="module"></script>
+  <script src="./js/data.js"></script>
 </body>
 
 </html>
