+++ js/game.js
@@ -1,0 +1,2 @@
+//модуль с демоверсией 
+
