+++ js/picture.js
@@ -1,0 +1,1 @@
+//найдем в документе шаблон котрый будем копировать
