+++ js/setup.js
@@ -1,0 +1,1 @@
+// модуль с настройкой
